bt node fixes
